requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.10
Django==5.1.3
django-crispy-forms==2.3
djangorestframework==3.15.2
pillow==11.0.0
sqlparse==0.5.2
typing_extensions==4.12.2
